1.结构大致理清

+++ ConsoleApplication2/base_serializable.h
@@ -3,7 +3,7 @@
 class base_serializable
 {
 public:
-	virtual size_t Serialize(std::vector<wchar_t>& Byte) = 0;
+	virtual int Serialize(std::vector<wchar_t>& Byte) = 0;
 	virtual void DeSerialize(std::vector<wchar_t>& ByteVssd,int &Pos) = 0;
 	base_serializable();
 	~base_serializable();
